fixed concepts, and fold expressions
