+++ Inc/C++Utilities/CppUtils.hpp
@@ -67,9 +67,10 @@ using std::unordered_map;
 using std::vector;
 using std::queue;
 using std::map;
+using std::stringstream;
 
-template<class Type>
-concept Integral = is_integral<Type>::value;
+// template<class Type>
+// concept Integral = is_integral<Type>::value;
 
-template<class Type>
-concept NotIntegral = !is_integral<Type>::value;
+// template<class Type>
+// concept NotIntegral = !is_integral<Type>::value;
